Rename Solution and Projects
