+++ ActiveDirectoryQuerier.Tests/ActiveDirectoryCommandsTests.cs
@@ -1,8 +1,8 @@
 ﻿using System.Collections.ObjectModel;
 using System.Management.Automation.Runspaces;
-using FAFB_PowerShell_Tool.PowerShell;
+using ActiveDirectoryQuerier.PowerShell;
 
-namespace FAFB_PowerShell_Tool.Tests;
+namespace ActiveDirectoryQuerier.Tests;
 
 public class ActiveDirectoryCommandsTests
 {
