+++ ActiveDirectoryQuerier.Tests/PowerShellExecutorTests.cs
@@ -1,7 +1,7 @@
 ﻿using System.Management.Automation.Runspaces;
-using FAFB_PowerShell_Tool.PowerShell;
+using ActiveDirectoryQuerier.PowerShell;
 
-namespace FAFB_PowerShell_Tool.Tests;
+namespace ActiveDirectoryQuerier.Tests;
 
 public class PowerShellExecutorTests
 {
