+++ ActiveDirectoryQuerier/App.xaml
@@ -1,7 +1,6 @@
-﻿<Application x:Class="FAFB_PowerShell_Tool.App"
+﻿<Application x:Class="ActiveDirectoryQuerier.App"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:local="clr-namespace:FAFB_PowerShell_Tool"
              StartupUri="MainWindow.xaml">
     <Application.Resources>
