+++ ActiveDirectoryQuerier/AssemblyInfo.cs
