+++ ActiveDirectoryQuerier/MainWindow.xaml
@@ -1,13 +1,13 @@
-﻿<Window x:Class="FAFB_PowerShell_Tool.MainWindow"
+﻿<Window x:Class="ActiveDirectoryQuerier.MainWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:FAFB_PowerShell_Tool"
+        xmlns:activeDirectoryQuerier="clr-namespace:ActiveDirectoryQuerier"
         mc:Ignorable="d"
         Title="Active Directory Querier" Height="720" Width="1280">
     <Window.DataContext>
-        <local:MainWindowViewModel />
+        <activeDirectoryQuerier:MainWindowViewModel />
     </Window.DataContext>
 
     <Grid>
